requirements.txt
Appium-Python-Client~=2.7.0
pytest~=7.1.3
selenium~=4.5.0
pytest-bdd~=6.0.1